predictive-service/requirements.txt
pandas
prophet
fastapi
uvicorn
fastapi
python-dotenv
psycopg2-binary
apscheduler
